+++ apx_sdk/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `verification_method_id` method to `DidPkh` type.
 - Added support for `@request-target` RFC-9421 derived component.
 - Added `without_query` method to `ApUri` type.
+- Added `is_same_uri` function to `url::canonical` module.
 
 ### Changed